Ensure that weights are floats on all functions in C API

+++ src/shared_library/thot.cc
@@ -645,7 +645,7 @@ void* llWeightUpdater_create()
 }
 
 void llWeightUpdater_updateClosedCorpus(void* llWeightUpdaterHandle,const char** references,const char*** nblists,const double*** scoreComps,const unsigned int* nblistLens,
-                                        double* weights,unsigned int numSents,unsigned int numWeights)
+                                        float* weights,unsigned int numSents,unsigned int numWeights)
 {
   LlWeightUpdaterInfo* llwuInfo=static_cast<LlWeightUpdaterInfo*>(llWeightUpdaterHandle);
